+++ source/LeLib/limage/multicolorimage.h
@@ -175,7 +175,7 @@ public:
 
    void ExportCompressed(QString f1, QString f2, QString f3) override;
 
-   virtual void CompressAndSave(QByteArray& chars, QVector<int>& screen, int x0,int x1, int y0, int y1, int& noChars, int compression, int maxChars);
+   virtual void CompressAndSave(QByteArray& chars, QVector<int>& screen, int x0,int x1, int y0, int y1, int& noChars, double compression, int maxChars);
 
    void SetColor(uchar col, uchar idx) override;
